requirements.txt
numpy>=1.19.0
Pillow>=8.0.0
torch>=2.0.0